crawling/requirements.txt
boto3
python-dotenv
requests
bs4
pandas